Move trainer setup to battle_setup.c
